Fix no default bug on stricter compilers
So msvc is a bit stricter about this than gcc, but when you
forcefully say "hey no copy constructor and move constructor" the
standard says the compiler *may* give you a default constructor.
Gcc does give you one, but msvc doesn't, so for us to be able to initialize a
PipelineConfigInfo on msvc we need a default constructor. Easy fix and was
preventing us from building on windows with msvc.